bq_streaming_cdc/pyproject.toml
[project]
name = "bq-streaming-cdc"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-cloud-bigquery>=3.40.0",
    "google-cloud-bigquery-storage>=2.36.0",
    "grpcio-tools>=1.76.0",
    "protobuf>=6.33.4",
]

[dependency-groups]
dev = [
    "ruff>=0.14.14",
]
